requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==25.4.0
blinker==1.9.0
boto3==1.42.70
botocore==1.42.70
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.5
exceptiongroup==1.3.1
filelock==3.19.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.83
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.0.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.33.6
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.12.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
s3transfer==0.16.0
six==1.17.0
smmap==5.0.3
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
